Assert integrand returns a real value for quadpack routines
Because the quadpack routines call coerce-float-fun to be able to
evaluate the integrand, the realpart is returned by default.  Instead,
we bind `*plot-realpart*` to nil so that evaluation can return complex
numbers.  Then modify float-integrand-or-lose to check for floats
since it was previously checking for cl:number.

Finally to make this all work, declaim `*plot-realpart*` to be special
because we compile plot.lisp (that defines it) **after** compiling
quadpack.lisp.  Reordering causes other problems.