Fix #4143: atan(0,minf) should be %pi.
Fix typo in atan2 simplifier where we were taking the realpart of x
instead of the realpart of y to handle the case of atan2(y,minf).  Add
a comment that we're making atan2 continuous with the second quadrant
along the branch cut.

Add a test for this case and update the documentation to say that
we're continuous with second quadrant on the branch cut.