Fix #4121: atan2(0, <nz>) doesn't simplify
atan2(0,x), when sign(x) is nz should simplify to %pi. The code was
checking for sign(x) being neg, not nz.  Modify the code to look for
both, just like how we simplify atan2(0,x) when sign(x) is pos, or pz.

Add test for this.  We also fix a typo where the start of comment was
inadvertently left off.