rtest_raddenest: fix a test and the comments
Due to the recent changes to the sign function, we no longer
incorrectly simplify sqrt((sqrt(1-sqrt(x))-1)^2) to
abs(sqrt(1-sqrt(x))-1) unless the argument in brackets is actually
real, so we have to add additional assumptions to the test.
Reformulating the test to use a plus in place of a minus sign makes
things a bit clearer.