Make bessel_k TeX formula match A&S and DLMF formulation
Minor tweak to the formula for bessel_k to make it match the formula
given by A&S and DLMF.  Since we give the references to the formulas,
we should make our formula look the same.  The existing formula was
equivalent, but no need to make it superficially different from the
references.