Fix [#4195] and [#4200]: cmucl-21e limit error
[#4200] mentions that cmucl (and clisp) warn about ignoring some
special variables that are arguments to `simplim%limit`.  To fix this,
we rename the offending arguments from `e` and `x` to `ee` and `xx`.
This isn't great, but I don't know where or how these variables are
special.

[#4195] is also fixed by this change.  The underlying issue there for
cmucl was that for some reason the call to `maxima-substitute` in
`stirling` seemed to cause problems.  The third arg is a constant
list, and it appears that it was getting modified.  This was verified
by consing a new value for the third arg each time.  The error no
long occurs.