Fix warning about ignoring special var
Both cmucl and clisp warn that the variable `A` is a special variable
that is being ignored.  To fix this, just rename the variable.

I am unable to find where `A` becomes a speical variable.  This only
seems to happen when compiling maxima from scratch.  If you compile
hypergeomtric.lisp by itself, there's no such warning.