Fix #4200:  Rename special variables
sprdet.lisp declares `x` as a special variable and then undeclares it
later.  However, `x` only appears to be used as a speical variable in
the function `nbn`.  So add `x` as a parameter to `nbn` and remove the
special declarations for `x`.

As a test, we use `x` again in `simplim%limit` again to see there are
compiler warnings/notes.  Neither clisp nor cmucl complain about
ignoring the special variable `x` any more.

Full testsuite (including share tests) pass.