Do not declare `e` to be special in sin.lisp
It seems that `e` does not need to be declare to be a special variable
in sin.lisp.  Removing the declaration does not cause any compiler
warnings/notes and the warning in limit.lisp about `e` being an
ignored special variable is gone now in clisp.

My guess is that there was some confusion because `e` is used to match
things via the m2 matcher.  But it's not an actual variable; it's used
as a simple symbol to denote the item that was matched.

Possibly `w` and `y` are also used in this way.  I didn't check.