Rename k to *k*
Make it more obvious that `k` is a special variable by renaming it to
`*k*`.

This should probably be reworked because `*k*` is set in `ratroot` and
is used only in `subst4`, which is only called from `subst41`, which
in turn is only called from `ratroot`.  It could be a lexical var by
passing the value of `*k*` to the necessary functions.