Rename *k* to k and make it lexical.
Remove the declare-top declaration for *k* and add `k` as the last
argument of `subst4` and `subst41`.  `subst4` referenced `*k*`, so it
needs the arg and `subst41` calls `subst4` so it needs the extra k arg
too.  `subst41` is only called by `ratroot` which defines and set the
value of `k`.