Rename exp to *exp*
It's a special variable so rename it.

Full testsuite (+share) passes.

Future work.  It almost works by removing the special declaration for
exp and adding it as an arg to substint and updating all the callers.
Just 4 integration tests fail.  I don't know why these fail.
Unfortunate; it would have been nice to get rid of the special var.