*expres* does not need to be special
Previously, it was referenced only in the function elem as a special.
Elem is a predicate for the schatchen matcher.

From reading the comments in schatc.lisp, the pattern for a variable
can have additional args for the predicate.  We take advantage of that
feature as add *expres* as the second arg to elem.  And update elem to
accept the second arg.

I think this is a pretty nice feature for the matcher because I think
there are lots of places where the predicate uses a special var for
matching stuff which could be replaced by adding it as an arg.

Full testsuite + share passes with cmucl.