Add comments on how we might be able to remove some specvars
Describe some ways to get rid of *powerl* and *ans*.  Untested, but
these look feasible.  Not sure if it's the best, but getting rid of
special variables seems like a big win in terms of readability and
reducing side-effects.  (Well, they are in the closure, but at least
we know exactly the functions that need it.)

One minor indentation change to make the code look better and easier
to read since it was indented incorrectly.