Make a closure for *powerl* for intform and integrator.
As mentioned in the notes, we can create a closure for *powerl* that
is used only by intform and integrator.  Then we don't need to
declare *powerl* to be special anymore, and it's scope is also limited
to just the functions that need it.

Full testsuite + share passes for clisp, cmucl, and ecl.