Make a closure for *ans* for integrate1 and sinint.
As mentioned in the notes, we can create a closure for *ans* that
is used only by integrate1 and sinint.  Then we don't need to
declare *ans* to be special anymore, and it's scope is also limited
to just the functions that need it.

Unfortunately doing this rearranges the code a little so stage I
doesn't appear first in the code.

Perhaps an alternative is to just declare *ans* special in
sinint (which sets it) and integrate1 which uses it?  The call tree is
complicated.

Full testsuite + share passes for cmucl.