*ans* does not need to be a special variable at all.
Rename *ans* to ans and comment out the special declaration for *ans*.

In integrate1, it was used to accumulate results and was (mostly)
returned as the value.  The only other place was in sinint that set it
as the result (mostly) of integrator and then passed that as the arg
to sum-of-intsp.  None of the uses require *ans* to be special.

Full testsuite + share passes with cmucl.