Use varp2 predicate instead of varp in some places.
Replace the varp predicate with varp2 predicate that accepts a second
variable for the var.

Work in progress, but we updated m2-exp-type-10-1, m2-ratrootform, and
m2-chebyform.  This also means updating all callers to pass in the
value of (specvar) var.

Full testsuite + share passes with cmucl.

There are more places to change.