Add freevar2 predicate like freevar but takes the var arg
freevar is used in quite a few other places, so define freevar2 like
freevar that takes an argument for the variable.  Replace as many uses
as we can of the freevar predicate with freevar2 predicate.

It helps also to have simple-trig-arg take a second arg for the
variable var so we can use freevar2 instead of freevar.  This requires
updating callers to pass in the value.  Hence, the update for
defint.lisp to pass in the var as well.

Finally add a note that the m2-a*x+b/c+d predicate doesn't seem to be
used anywhere in Maxima, as shown by a git grep of the repo.

Full testsuite + share passes with cmucl.