Make dependency of elem on var explicit by passing in the arg
elem uses freevar, but we want to use freevar2 to make the dependency
on var explicit.  Hence, add a third arg (named var2) to elem to hold
the value of var.  Then we can use freevar2.

This means we need to update the one place where elem is used to pass
in the value of var.

Full testsuite + share passes with cmucl.