Use FREEVAR2 in SUBST4
From the call chain, RATROOT calls SUBST41 with the B arg equal to VAR
and SUBST41 calls SUBST4.  SUBST41 sets ROOTVAR equal to B and is
therefore equal to VAR.  So instead of using FREEVAR (that references
the specvar VAR), we can use FREEVAR2 with a second arg of
ROOTVAR (VAR).

This gets ride of the implicit reference to the special variable VAR
in SUBST4.

Full testsuite + share passes with cmucl.