Use freevar2 instead of freevar in supertrig
This fixes the last use of freevar in sin.lisp (except for freevar
itself).  To do this, add a new arg `var2` to `supertrig` to hold the
variable of integration.  Then the single caller `monstertrig` needs
to updated to pass in the additional arg to `supertrig`.
`monstertrig` already takes a `var` arg, so we're set.

Full testsuite + share passes with cmucl.