Replace last use of varp with varp2
The last use of `varp` is in `checkderiv1`.  To replace it with
`varp2`, we need to `checkderiv1` to take an extra arg for the value
of `var` (called `var2` so we don't use the special var `var`
unnecessarily).  This means `checkderiv` needs an extra arg too.
`checkderiv` is only called from `intform`.

Full testsuite + share passes with cmucl.