Rename var to var2 in chebyf
Make sure we don't need the special var `var` by renaming `var` to
`var2` in `chebyf`.  This also requires changing some of the lists
that are arguments to subliss to use backquoted expressions so we can
substitute the value of `var2` into the expressions.

Full testsuite + share passes with cmucl.