Rename var to var2 in integrate-exp-special and add freevar02
Make sure we don't need the special var `var` by renaming `var` to
`var2` in `integrate-exp-special`.

However, for this to work, we also need to add `freevar02` that is
like `freevar0` (defined in hypgeo.lisp) but takes a second arg for
the variable.  And then replace all uses of `freevar0` with
`freevar02`.  All of these are in the functions with prefix
`m2-exp-type-` used by `integrate-exp-special`.

Full testsuite + share passes with cmucl.