Rename var to var2 in sinint and add free12
Make sure we don't need the special var `var` by renaming `var` to
`var2` in `sinint`.

However, for this to work, we also need to add `free12` that is
like `free1` (defined in schatc.lisp) but takes a second arg for
the variable.  And then replace almost all uses of `free1` with
`free12`.  The places where we can't are functions that still
implicitly reference `var` that we have not updated.

Full testsuite + share passes with cmucl.