Rename *trigarg* to trigarg and pass it around as needed.
The special var `*trigarg*` can be removed.  It's only accessed in
`supertrig` which we can fix by adding as a third arg to `supertrig`.
`monstertrig` calls `supertrig` so modify calls to pass the value of
`trigarg` to `supertrig`.  And rename `*trigarg*` to `trigarg` in
`monstertrig`.

Now `*trigarg*` doesn't need to be special.

Full testsuite + share passes with cmucl.