Rename *exp* to expr in sinint
`sinint` receives the `*exp*` arg and mostly passes it on to other
routines.  It does modify `*exp*`, but it seems the modification is
is only used as args to other functions.  It looks like no other
called functions depend on the modification in a hidden fashion.

Full testsuite + share passes with cmucl.