Remove global decl of *exp* and use function-specific declaration
Instead of doing a declare-top for `*exp*`, just add special
declarations for it in `intform` and `integrator` which are the only
two places left that use `*exp*`

I tried various things to get rid of `*exp*`:

o Have the functions close over `*exp*` the same way they close over
  `powerl`.  This fails with integrals not being computed correctly.
o Add `*exp*` as a parameter to `intform`.  This also fails.

I don't know the neither of these approaches fail.

One such failure was integrate(sqrt(2-2*cos(x)),x).  The result is
completely wrong.

With this change, the testsuite + share tests pass with cmucl.