Remove specvars in powerlist
In `powerlist`, there were the special variables `*b*`, `*c*`, and
`*d*`.  These don't need to be special and can be replaced by
lexicals.  This works because all the calls to powerlist (in
`integrator`) return the value of `powerlist` immediately.  Hence the
modifications to these special vars are never used.

Also, `subst10` uses `*d*`.  This can be fixed by adding a third arg
to subst10 for the value of `*d*`.  This works because `subst10` is
only called by `powerlist`.

Full testsuite + share passes with cmucl.