Remove specvar *b1* from rat1 and rats
`rat1` set the specvar `*b1*` and called `rats` which used the value
of `*b1*`.  There are no other uses of `*b1*` and `rats` is only
called from `rat1`.  Thus, rename `*b1*` to `b1` and add second arg to
`rats` for the value of `b1`.

Full testsuite + share passes with cmucl.