Replace specvars *a* and *b* in rat1
`rat1` used the specvar `*b`* and called `rats` that used the specvar
`*a*`.  These can be replaced by lexicals by passing the values of
`*a*` and `*b*` to `rat1` and `*a*` to `rats`.

Thus, rename `*a*` and `*b*` to `aa` and `bb`, respectively.

This requires updating all calls to `rat1` to pass in these extra
values.  `rat1` is only called in `trigint` (which initializes `*a*`
and `*b*`).  `rat1` is a predicate to m2, so we can have pass in the
extra values there.  There's also a one plain call to `rat1` that
needs updating.

Full testuite + share passes with cmucl.