Update m2 calls to use the value of aa and bb
Because `trigint` used to communicate with `rat1 with the special
variables `*a*` and `*b*`, `trigint` set the values of `*a*` and `*b*`
for use by `rat1`.  Now that `rat` takes args for these values, we can
get rid of these variables and use the values directly in the call to
`rat1`.

Full testsuite + share passes with cmucl.