Get rid of *c* in trigint, odd1, and f3.
`f3` uses `*c*` and is only called from `rats`, which is only called
from `rat1`.  We can replace the use of `*c*` by passing in an
additional arg to `rat1`.

`odd1` uses `*c*` and is only used as a m2 predicate in `trigint`.  By
adding a second arg to `odd1`, we can pass in the value of `*c*`
instead.  And instead of setting `*c*` in `trigint`, we just use the
fixed value in the predicates.

Full testsuite + share passes with cmucl.