Support abcl unicode characters for display
Define a function `get-unicode-char` that takes a keyword representing
the Unicode character that we want and return the corresponding
Unicode character.

We do it this way because there's no guarantee that all Lisps use the
same name for the character.  (In particular ABCL doesn't have these
names.)

`get-unicode-char` has an alist between the keyword and the code of
the corresponding character.  The keyword is the name of the Unicode
character except that underscores are replaced by dashes (in basic
lisp style to use dashes.)