Add arg to hyp-atanh so it doesn't access specvar var
Doing this mean all the callers need to be updated.  Which means
`simp2f1` and `dispach-spec-simp` also need the additional arg to hold
the variable, which means their callers need updating too.

We name this extra arg `arg` because `$hgfred` has such an arg, and it
binds the specvar `var` to `arg` at the beginning.

Full testsuite + share passes with cmucl.