Add arg to step4-int so it doesn't access specvar var
Doing this mean all the callers need to be updated.  Which means
`step4` also need the additional arg to hold the variable, which means
their callers need updating too.

We name this extra arg `arg` because `$hgfred` has such an arg, and it
binds the specvar `var` to `arg` at the beginning.

Also forgot to rename `var` to `arg` in `simp2f1` in the previous
commit.  Do it now.

Full testsuite + share passes with cmucl.