Rename var to arg in confl and add arg to kummer
AFAICT, `var` in `confl`, doesn't really need to be special so rename
it to `arg`.  While we're at it, also add a second arg to kummer so
that it doesn't need to access the specvar `var`.  This also
simplifies the code slightly in `confl` where we were binding `var` to
`z`.  We can just pass `z` to `kummer`.

Full testsuite + share passes with cmucl.