Add arg to trig-lot-1 so it doesn't access specvar var
All occurrences of `var` in `trig-log-1` need to be renamed to `arg`.
This requires updating `trig-log` to have an extra arg too.  And all
callers need to be updated as well.

We name this extra arg `arg` because `$hgfred` has such an arg, and it
binds the specvar `var` to `arg` at the beginning.

Full testsuite + share passes with cmucl.