Add arg to trig-log-3a-exec so it doesn't access specvar var
All occurrences of `var` in `trig-log-3a-exec` need to be renamed to
`arg`.  This requires updating `trig-log-3` to have an extra arg too.
And all callers need to be updated as well.

We name this extra arg `arg` because `$hgfred` has such an arg, and it
binds the specvar `var` to `arg` at the beginning.

Full testsuite + share passes with cmucl.