Add arg to trig-sin so it doesn't access specvar var
All occurrences of `var` in `trig-sin` need to be renamed to
`arg`.  And all callers need to be updated as well.

We name this extra arg `arg` because `$hgfred` has such an arg, and it
binds the specvar `var` to `arg` at the beginning.

Full testsuite + share passes with cmucl.