Add arg to simp2f2 so it doesn't access specvar var
All occurrences of `var` in `simp2f2` need to be renamed to `arg`.
This also requires adding the arg to `step7`.  And all callers need to
be updated as well.

We also forgot to rename `var` in a few places in `dispatch-spec-simp`
so do it now.

Full testsuite + share passes with cmucl.