Add arg to create-any-poly so it doesn't access specvar var
All occurrences of `var` in `create-any-poly` need to be renamed to
`arg`.  And all callers need to be updated as well.  This also
required adding `arg` to `create-poly`, `simpg-exec`, and `simpg` and
updating the callers as well.

In `$hgfpoly`, we can get rid of the code that binds `var` to `arg`
because it only calls `create-any-poly` that now takes an arg for the
value of `arg`.

Full testsuite + share passes with cmucl.