Make optional arg to hyp-cos required
`hyp-cos` had an optional arg that was initialized to `var` by
default.  Replace this optional arg with a required arg.  We leave the
name of the arg as `z` since that's what the code uses.

Update callers to `hyp-cos` to always specify the arg.

Full testsuite + share passes with cmucl.