Rename var to var2 in the third set of pattern matchers
`var` is a special variable in this code but doesn't need to be.
Rename `var` to `var2` to break the connection to the special
variable.  We do all of the third set of pattern matchers from
`m2-a*t` to `m2-a*x^m+c`.  This is all of the m2 pattern matchers.

Full testsuite + share passes with cmucl.