Add arg to exp-converg-form for the variable
To break the connection to the specvar `var`, add `arg` to
`exp-converg-form` and also rename all uses of `var` to `arg`.  Update
callers to pass in the extra value.  This also requires adding an arg
to `special-convergent-formp` and updating callers.

We also missed changing the call to `sc-converg-form` in
`special-convergent-formp` to add the extra arg that we forgot to do
in the previous commit.

Full testsuite + share passes with cmucl.