Fix warnings about undefined vars for unicode chars
Move the defvars for the unicode box characters earlier in the file so
we don't get warnings about undefined variables.

Also fix one warning from `d-box-unicode` that `dmstr` is unknown and
assumed to be global.  We just need to make `dmstr` an `&aux` variable
like the other functions that use `dmstr`.  (Should we have just used
a `let` instead?)

The warnings are now gone.