Remove adding ccl character names and some doc changes
Since we don't need names for the unicode characters for drawing
boxes (and related chars), get rid of the code for ccl that added the
character names.

Minor indentation change for adding the feature
`:lisp-unicode-capable` to make it easier to see we're pushing stuff.

Finally, fix comment that said `maxima-unicode-enabled`.  I think's
it's `display2d-unicode-enabled` now.