Partially undo change to xtorterm and rename var1 to arg
Since we've determined that all calls to `xtorterm` pass `var` as the
second arg (`var1`), remove the recently added third arg and rename
`var1` to `arg`.  Update callers to pass only two values and verify
that the second arg is always `var` or equivalent.

Full testsuite + share passes with cmucl.