Add arg to funclogor%e to remove reference to specvar var
Add a parameter, named `ivar`, to `funclogor%e` to hold the variable
of integration (usually in the specvar `var`).  Rename `var` to `ivar`
in `dintbypart`.  We use `ivar` because this function already has a
local named `arg`.

Update callers appropriately.

Fix typo in dintbypart where we didn't change one use of `var` to
`var1`.

Full testsuite + share passes with cmucl.