Remove extra declaration of var and add special decl
There's an special declaration of `var` in the middle of the file
which is redundant because `var` is declared special at the top of the
file.

But also add a local special declaration for `var` for the call to
`polelist` in `rectzto%pi2` to make it clear, especially if we succeed
in removing the special var `var` everywhere else.

Full testsuite + share passes with cmucl.