Rename var to ivar in unitcir and zto%pi2
However, `unitcir` calls `numden` and `res`, both of which reference
the special variable `var`.  So in `unitcir`, bind `var` to
`ivar` (and declare `var` special, just in case) so that those two
functions get the information they need.  When (ir?) `numden` and
`res` get updated not te reference `var`, we can update this.

Full testsuite + share passes with cmucl.