Rename var to ivar in sin-cos-intsubs
To break the connection to the specvar `var`, rename it to `ivar` in
`sin-cos-intsubs`.  However, this means passing `ivar` to
`sin-cos-intsubs1` by adding `ivar` as a new arg for
`sin-cos-intsubs1`.  But that's not enough because this function calls
`pdis` that references `var`.  So we need to bind `var` to `ivar` for
the call to `pdis`.  When (if?) `pdis` gets updated not to reference
the specvar `var`, we can fix this.

Full testsuite + share passes with cmucl.