Add arg to remove access to specvar var for ggrm
`ggrm` access the specvar `var`.  Fix this by adding a new arg,
`ivar` for the value of `var`.  Rename `var` to `ivar` and update all
callers appropriately.

Also rename `var` to `ivar` in `keyhole`.  However, `keyhole` calls
`res` which references `var`, so we need to bind `ivar` to `var` for
`res` to work.  Do this with a simple let binding and add a special
declaration for `var`.

Full testsuite + share passes with cmucl.