numden and snumden needs var to be bound
In many places where `numden` and `snumden` are called, we don't have
`var bound (in the function).  But both reference the specvar `var`,
should we need to make sure the calls have `var` bound appropriately.

AFAICT, there are no tests that fail without this, but that could be
by accident where `var` was bound to the correct value farther up in
the call tree.  If (when?) we remove all references to the specvar in
defint.lisp, then these would probably fail.  But it's not clear if
there are any tests that would detect this.

Full testsuite + share tests pass with cmucl.