Add arg to take-principal and rename var in principal-value-integral
`take-principal` accesses the specvar `var`.  Fix this by adding a new
arg, `ivar` for the value of `var`.  Rename `var` to `ivar` and update
all callers appropriately.

In `principal-value-integral`, rename `var` to `ivar` to break the
connection to the specvar `var`.  This is a natural consequence of
adding an arg to `take-principal`.

Full testsuite + share passes with cmucl.