Add ivar to ztorat and mtorat
Both of these reference the specvar `var`.  Add an arg for this
instead to break the hidden dependency.  However, for this to work, we
need to modify all the places that say `#'mtorat` and `#'ztorat` to be
a lambda to capture the variable `ivar`.

Also add a comment in `difapply` that was testing if `fn1` was `eq` to
`mtorat`.  Even before this change, this was never true; no callers
ever passed the symbol `mtorat`; it was always `#'mtorat`.

Full testsuite + share passes with cmucl.