Define snumden-var and res-var to make dependency explicit
`snumden-var` takes an arg for the variable of interest.  Currently
just binds `var` to the arg and calls `snumden`.  This is a workaround
until we actually fix `numden` not to use the special variable.

Likewise define `res-var` to bind `var` before calling `res` the
references the special var `var`.

Full testsuite + share pass with cmucl.