Define resprog0-var and polelist-var to make dependency explicit
Define these functions to make dependency on `var` explicit.  But, as
we did for other functions, we just bind `var` to the arg and call the
real function.   This is a workaround until these functions make the
dependency explicit.

Also, we were binding `var` around a call to `pdis`.  That doesn't
seem necessary and was a mistake?  AFAICT, `pdis` doesn't actually
reference `var`.

Full testsuite + share pass with cmucl.