Add replacement for tansc that refs specvar var
The function `tansc` references the special var `var` so we add a
replacement `tansc-var` that takes an extra arg, `ivar`, and
binds `var` to it before calling `tansc`.  This is a workaround
until we fix `tansc` to make the dependency explicit.

(`tansc` calls `involve` that references `var`.)

Update all uses of `tansc` to use `tansc-var`.

Full testsuite + share passes with cmucl.