Add replacement for oscip that refs specvar var
The function `oscip` references the special var `var` so we add a
replacement `oscip-var` that takes an extra arg, `ivar`, and binds
`var` to it before calling `oscip`.  This is a workaround until we fix
`oscip` to make the dependency explicit.  (`oscip` calls `involve`.)

Update all uses of `oscip`.

Full testsuite + share passes with cmucl.