Add replacement for residue that refs specvar var
The function `residue` references the special var `var` so we add a
replacement `residue-var` that takes an extra arg, `ivar`, and binds
`var` to it before calling `residue`.  This is a workaround until we fix
`residue` to make the dependency explicit.

Update all uses of `residue`.

Full testsuite + share passes with cmucl.