Add replacement for deg that refs specvar var
The function `deg` references the special var `var` so we add a
replacement `deg-var` that takes an extra arg, `ivar`, and binds
`var` to it before calling `deg`.  This is a workaround until we fix
`deg` to make the dependency explicit.

Update all uses of `deg`.  This requires passing in `ivar` in
`difapply` and `diffhk`.  Update all callers appropriately.

Full testsuite + share passes with cmucl.