Add replacement for res1 that refs specvar var
The function `res1` references the special var `var` so we add a
replacement `res1-var` that takes an extra arg, `ivar`, and binds
`var` to it before calling `res`.  This is a workaround until we fix
`res1` to make the dependency explicit.

Update all uses of `res1`.

Full testsuite + share passes with cmucl.