Add replacement for polyp that refs specvar var
The function `polyp` references the special var `var` so we add a
replacement `polyp-var` that takes an extra arg, `ivar`, and binds
`var` to it before calling `res`.  This is a workaround until we fix
`polyp` to make the dependency explicit.

Update all uses of `polyp`.

Full testsuite + share passes with cmucl.