Bind var for plog in ztorat
In `ztorat`, we multiply the numerator by plog(-x).  In doing so, we
end up calling the plog simplifier that references `var`:

```
	   ((among var x)	;This is used in DEFINT. 1/19/81. -JIM
	    (return (give-up))))
```

Thus, we need to bind `var` before calling the simplifier for plog.
This is important.  Without this, we get failures in the testsuite.
With this, there's no problem.

More debugging needed.