Bind var for plog in ztorat
The simplifier for plog references the specvar `var` and it's
important to do the binding here in the call to `keyhole` in `ztorat`.
I guess the intent is to prevent the simplifier from converting
plog(-x) to log(x)+%i*%pi when x > 0, as happens in `ztorat`.

Full testsuite + share tests pass with cmucl.