Fix compiler warning about unknown variable VAR
`extra-simp` references the special variable `var`, but `var` was
declared unspecial earlier in the file.  Let's leave the unspecial
declaration as is, and declare `var` special in `extra-simp` to get
rid of the warning.

We should probably pass `var` in to remove that dependency.  For
another day, perhaps.

This occurs only with lisps like clisp and cmucl that actually support
the unspecial declaration.  I think sbcl doesn't support unspecial.