Remove declare-top declaration for var
The special variable `var` has been replaced by a lexical in almost
all places in defint.  We don't need the `declare-top` declaration for
`var` anymore.  Only a few places still need to reference it:

o The replacement functions that accept a lexical to bind `var` before
  calling the underlying function that references `var`.
o The one call to `keyhole` that requires `plog` not to simplify the
  arg until later.

In all of these cases, we locally declare `var` as special.  Hence the
file declaration is not needed anymore.

Full testsuite + share passes with clisp (2.49.93+ (2018-02-18)),
ecl (23.9.9), and cmucl (snapshot-2023-08).