Properly implement replacements to make dependency explicit
We had defined some replacement functions to make the dependency on
`var` explicit.  But these just added an arg to the functions and then
just called the originals.  Now it's time to implement the properly
without referencing the special vars.

Functions done:  %einvolve-var, involve-var, notinvolve-var,
tansc-var, oscip-var

Also, replace `polyp-var` with `polyp2`, which we implemented a while
a go and forgot about.

Full testsuite + share passes with cmucl.