Implement residue-var
To make it work, we also had to add `resm0-var` and `resm1-var` which also
had hidden references to the specvar `var`.

For consistency to how we modified other residu functions, the
variable is first in the arg list.  This requires updating defint to
match.

Full testsuite + share passes with cmucl.