Make *zd* lexical instead of special
To do this, `maybpc` returns `*zd*` as a second value instead of
setting it.  Then update `ggr1` that calls `maybpc` to accept the
second value.  And also return it as a second value.

Finally update `ggr` that calls `ggr1` to accept the second value.

Also, removed the unused functions logcpi and sumi that we changed in
the previous commit.

Full testsuite + share passes with cmucl.