Make rlm* lexical
`rlm*` doesn't need to be special.  It's initialized in `logcpi0`,
modified there and used in forming the result.  It's not referenced
anywhere else.  Hence, there's no reason for it to be special.  This
also means we don't need to bind it to nil in `log*rat`.

Full testsuite + share passes with cmucl.