Make plm* a lexical instead of special
`plm*` is initialized in `log*rat` and the value is set in `logcpi0`.

To make it lexical, have `logcpi0` return `plm*` as the second value.
This is called by `logcpi` and returns immediately, so `logcpi` will
return 2 values too in this case.  In `log*rat`, update the call to
`logcpi` to accept the two return values, updating the array and
`plm*` appropriately.  We also call `logcpi` when the `dotimes` loop
is done, but we don't need to do anything with the second value.

`logcpj` references `plm*`, so pass it as an additional arg.

Full testsuite + share passes with cmucl.