Fix error in previous commit confusing pl and pl*
In renaming `pl*` to `pl`, we got confused because `logcpi0` already
has a lexical named `pl`.  Surprisingly, the testsuite passes with
this error, but we should fix it.  `pl` was supposed to be the result
from `polelist-var`.  So name it `polelist` instead of `pl`, leaving
the renaming of `pl*` to `pl`.

While we're at it fix a compiler warning about unused vars `pl*`,
`rl*`, `pl*1`, and `rl*1` in `log*rat`.  We made these lexical, so we
don't need these anymore.

Full testsuite + share passes with cmucl.