Windows: Add (basic) support for CCL64 in the installer.
One can enable this with the CMake option "-DWITH_CCL64=YES"
when building the installer.

Some small changes to the build system were necessary,
especially changing the pipe to --eval commands for CCL64.
(don't know, why, but piping the commands to CCL64 (using
the windows version with wine as emulator) did not work.

But probably the --eval commands are the 'cleaner' method
anyways (and maybe that should be changed for other Lisp
compilers too?).