Tried to support CCL (32 bit) for Windows. Does not work. Documented that.
This does not work most likely because a compiler bug in ccl(32).

So I am not commiting the changes, it does not help, if it works
'in theory'. Maybe I can check again, if CCL make their next release.
(Let's see, if they will include a 32 bit version at all, they are
considering 32 bit support close to obsolete...)