Fix #4265: integrate(sin(n*x)*sin(m*x),x,0,2*%pi) is always 0
integrate(sin(n*x)*sin(m*x),x,0,2*%pi) is 0 except when n=m.  To fix
this, we ask the user if n=m.  We also handle the integrands
cos(n*x)*cos(m*x) and cos(n*x)*sin(m*x).

However, this breaks abs_integrate when it is loaded.  So, if it is
loaded, we don't ask the user in these cases.

The full testsuite including share tests pass.