Add special case for maxima-init.lisp and maxima-init.mac
When a cindex was added for "maxima-init.lisp" and "maxima-init.mac"
to describe these files, the html index builder was converting the "-"
to " ".  This means that the html help can't find these
entries.  (When maxima starts, there's a warning about these.)

For these entries, we don't want "-" to be replaced.  To fix this, add
these entries to `handle-special-cases` which is meant for fixing up
these kinds of mistakes.

With this, starting maxima no longer warns about missing entries.