Allow display2d_unicode to accept only true and false
Use the `:setting-predicate` option for `defmvar` to only allow `true`
and `false` to be assigned to `display2d_unicode`.

But for lisps that do not support Unicode, setting `display2d_unicode`
to any value other than `false` causes an error with a message saying
it's not supported for this Lisp.