Fix undefined function error in TANSC1-V
in `tansc-var`, there's a labels function `tansc1-v`.  In this
function we call `(mapcar 'tansc1-v (cdr e))`.  This causes an error
about an undefined function `tansc1-v`.  To fix this, change
`'tansc1-v` to `#'tansc1-v`.  This no longer produces an error.

This was found trying to evaluate
`integrate(log(sec(x)^2+tan(x)^4),x,0,%pi/2)`.  Now we get a noun form
instead of an error.

Noticed this in cmucl and ccl64.