Implement jacobi_am function
Implement Jacobi amplitude function for complex values of z and m.

This uses the AGM algorithm to compute the result, based on code by
Dimiter Prodanov from the mailing list, 2024/04/22.

This replaces the old jacobi-am code (which hasn't been deleted yet).
This currently only produces the correct results for real z and real m
with |m| <= 1.

More work is needed.

There's also an attempt to use the q-series to evaluate jacobi_am.
This also seems to work only for real z and real m with |m| < 1.  Not
sure which approach is better.