Allow :TYPE to create an 'assign function
To support allowing :type to create an 'assign function, we need to
change the value of the :type option.  It is now a list of the form
`(type-spec maxims-desc)` where `type-spec` is a Lisp type-specifier
suitable for `typep`.  `maxima-desc` is a string that is used in the
error message to describe better what the expected type was instead of
using the Lisp type-specifier.  (The Lisp type-specifier is probably
no useful for most people who aren't familiar with Lisp types.)

Then `:type` will create a function to check if the value is a type
that matches the type-specifier.  However, this is only done if the
user hasn't specified a `:setting-predicate`, :setting-list` or a
`:properties` with an `'assign` property.

But in any case the variable is declared to have the given type.

For debugging, we now produce a deprecation message if a `defmvar`
includes `fixnum`, `boolean`, `string`, or `flonum`.  This is
temporary.