Remove old code that memoized bf values
Remove the defmacro for memoizing functions.  It was converted to a
macrolet, plus that defmacro was buggy.  The original hand-written
code for memoizing each function was superseded by the macrolet so we
can remove that too, which was already commented out.

Forgot to remove these when the issue was fixed.